Add readme

+++ extractMaterialsModelsFromVMF.py
@@ -3,26 +3,45 @@ import os
 import glob
 import shutil
 import string
+import sys
 from termcolor import colored
 
-vmfName = 'rp_expecto_day_vz4.vmf'
+print('Starting the process..')
+
+# name of the VMF file that will be processed
+# it must be in the same folder with this .py file
+vmfName = 'example.vmf'
+
+# path of Garry's Mod\garrysmod
+# be sure that you actually put Garry's Mod\garrysmod not only Garry's Mod\
 garrysmodGarrysmodPath = "D:\Program Files (x86)\SteamLibrary\steamapps\common\GarrysMod\garrysmod"
 
 vmfFileName, vmfExtension = os.path.splitext(vmfName)
 
-with open(vmfName) as vmfFile:
-    vmfContent = vmfFile.read()
+try:
+    with open(vmfName) as vmfFile:
+        vmfContent = vmfFile.read()
+except:
+    print(colored('Error while reading VMF file. Please be sure that the path is correct.', 'red'))
+    sys.exit()
+
+if not os.path.isdir(garrysmodGarrysmodPath):
+    print(colored("Error while reading Garry's Mod path. Please be sure that it looks like this:", 'red'),
+          "...steamapps\\common\\GarrysMod\\garrysmod")
+    sys.exit()
+
+print('VMF file has been read, total length:', colored(len(vmfContent), 'yellow'))
 
 materials = re.findall(r"\"material\" \"(.+)\"", vmfContent)
 materials = set(materials)
-materials = {"materials" + (material[0]=='/' and material or '/'+material).lower() + ".vmt" for material in materials}
-materials = {material.replace('/','\\') for material in materials}
+materials = {"materials" + (material[0] == '/' and material or '/' + material).lower() + ".vmt" for material in materials}
+materials = {material.replace('/', '\\') for material in materials}
 
 models = re.findall(r"\"model\" \"(.+)\"", vmfContent)
 models = set(models)
-models = {model.replace('/','\\') for model in models}
+models = {model.replace('/', '\\') for model in models}
 
-contentsDistPath = os.path.join(os.getcwd(),vmfFileName+"_map_contents\\")
+contentsDistPath = os.path.join(os.getcwd(), vmfFileName + "_map_contents\\")
 
 if os.path.isdir(contentsDistPath):
     shutil.rmtree(contentsDistPath)
@@ -30,6 +49,7 @@ if os.path.isdir(contentsDistPath):
 os.mkdir(contentsDistPath)
 
 # https://stackoverflow.com/a/17197027
+
 def strings(filename, min=1):
     strs = []
 
@@ -44,18 +64,19 @@ def strings(filename, min=1):
             result = ""
         if len(result) >= min:
             strs.append(result)
-    
+
     return strs
+
 
 def findVtfPathsFromVmt(vmtPath):
     with open(vmtPath) as materialFile:
         materialFileContents = materialFile.read()
 
     possibleVtfFileNames = re.findall(r"\"([^%$ {[\n,]+?)\"", materialFileContents)
-    possibleVtfFileNames = {vtfPath for vtfPath in possibleVtfFileNames }
-    possibleVtfFileNames = {vtfPath.replace('/','\\').lower() for vtfPath in possibleVtfFileNames}
-    possibleVtfFileNames = {vtfPath.endswith('.vtf') and vtfPath or vtfPath+".vtf" for vtfPath in possibleVtfFileNames}
-    
+    possibleVtfFileNames = {vtfPath for vtfPath in possibleVtfFileNames}
+    possibleVtfFileNames = {vtfPath.replace('/', '\\').lower() for vtfPath in possibleVtfFileNames}
+    possibleVtfFileNames = {vtfPath.endswith('.vtf') and vtfPath or vtfPath + ".vtf" for vtfPath in possibleVtfFileNames}
+
     paths = set()
 
     for name in possibleVtfFileNames:
@@ -66,13 +87,19 @@ def findVtfPathsFromVmt(vmtPath):
 
     return paths
 
+totalExtractedCount = 0
+
 def copyFileWithDirs(filePath):
+    global totalExtractedCount
+
     targetPath = os.path.join(contentsDistPath, os.path.dirname(os.path.relpath(filePath, garrysmodGarrysmodPath)))
     os.makedirs(targetPath, exist_ok=True)
     shutil.copy(filePath, targetPath)
+    totalExtractedCount = totalExtractedCount + 1
+
 
 for material in materials:
-    sourcePath = os.path.join(garrysmodGarrysmodPath,material)
+    sourcePath = os.path.join(garrysmodGarrysmodPath, material)
 
     if os.path.isfile(sourcePath):
         copyFileWithDirs(sourcePath)
@@ -80,19 +107,20 @@ for material in materials:
         for path in findVtfPathsFromVmt(sourcePath):
             copyFileWithDirs(path)
 
+
 for model in models:
-    sourcePath = os.path.join(garrysmodGarrysmodPath,model)
-    
+    sourcePath = os.path.join(garrysmodGarrysmodPath, model)
+
     if os.path.isfile(sourcePath):
         strs = strings(sourcePath)
-        materialPath = strs[-1].replace('/','\\')
+        materialPath = strs[-1].replace('/', '\\')
         materialPath = materialPath.lower()
-        materialPath = os.path.join('materials',materialPath)
-        materialRealPath = os.path.join(garrysmodGarrysmodPath,materialPath)
+        materialPath = os.path.join('materials', materialPath)
+        materialRealPath = os.path.join(garrysmodGarrysmodPath, materialPath)
 
         if (os.path.isdir(materialRealPath)):
             for possibleMaterialName in strs[1:-1]:
-                possibleMaterialRealPath = os.path.join(materialRealPath,possibleMaterialName+'.vmt')
+                possibleMaterialRealPath = os.path.join(materialRealPath, possibleMaterialName + '.vmt')
                 if (os.path.isfile(possibleMaterialRealPath)):
                     copyFileWithDirs(possibleMaterialRealPath)
                     for path in findVtfPathsFromVmt(possibleMaterialRealPath):
@@ -100,7 +128,9 @@ for model in models:
         else:
             print(colored("MATERIAL PATH NOT FOUND", 'red'), materialPath)
 
-        for file in glob.glob(sourcePath[:-3]+'*'):
+        for file in glob.glob(sourcePath[:-3] + '*'):
             copyFileWithDirs(file)
     else:
         print(colored("MODEL NOT FOUND", 'red'), model)
+
+print(colored('SUCCESS','green'),"A total of ",colored(totalExtractedCount, 'green')," files has been extracted to ",colored(contentsDistPath, 'green'))
